extract StepMatcher from PopupMenuFindStepActionDelegate
