+++ cucumber.eclipse.editor/src/main/java/cucumber/eclipse/editor/editors/PopupMenuFindStepActionDelegate.java
@@ -1,15 +1,10 @@
 package cucumber.eclipse.editor.editors;
 
-import gherkin.I18n;
-
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
-import java.util.regex.PatternSyntaxException;
 
 import org.eclipse.core.commands.AbstractHandler;
 import org.eclipse.core.commands.ExecutionEvent;
@@ -40,7 +35,6 @@ import cucumber.eclipse.steps.integration.Step;
 public class PopupMenuFindStepActionDelegate extends AbstractHandler {
 
 	private final static String EXTENSION_POINT_STEPDEFINITIONS_ID = "cucumber.eclipse.steps.integration";
-	private Pattern variablePattern = Pattern.compile("<([^>]+)>");
 
 	private List<IStepDefinitions> getStepDefinitions() {
 		List<IStepDefinitions> stepDefs = new ArrayList<IStepDefinitions>();
@@ -80,7 +74,7 @@ public class PopupMenuFindStepActionDelegate extends AbstractHandler {
 		String selectedLine = getSelectedLine(editorPart);
 		String language = getDocumentLanguage(editorPart);
 
-		Step matchedStep = matchSteps(language, steps, selectedLine);
+		Step matchedStep = new StepMatcher().matchSteps(language, steps, selectedLine);
 		try {
 			if (matchedStep != null) openEditor(matchedStep);
 		} catch (CoreException e) {
@@ -89,31 +83,6 @@ public class PopupMenuFindStepActionDelegate extends AbstractHandler {
 		return null;
 	}
 
-	Step matchSteps(String languageCode, Set<Step> steps, String currentLine) {
-		Pattern cukePattern = getLanguageKeyWordMatcher(languageCode);
-		
-		if(cukePattern == null)
-			return null;
-		
-		Matcher matcher = cukePattern.matcher(currentLine);
-		if (matcher.matches()) {
-			String cukeStep = matcher.group(1);
-		
-			// FIXME: Replace variables with 0 for now to allow them to match steps
-			// Should really read the whole scenario outline and sub in the first scenario
-			Matcher variableMatcher = variablePattern.matcher(cukeStep);
-			cukeStep = variableMatcher.replaceAll("0");
-			
-			for(Step step: steps) {
-				if (step.matches(cukeStep)) {
-					return step;
-				}
-			}
-			
-		} 
-		return null;
-	}
-	
 	private void openEditor(Step step) throws CoreException {
 		   IResource file = step.getSource();
 		   IWorkbenchPage page = PlatformUI.getWorkbench().getActiveWorkbenchWindow().getActivePage();
@@ -164,34 +133,6 @@ public class PopupMenuFindStepActionDelegate extends AbstractHandler {
 			return stepLine;
 		} catch (BadLocationException e) {
 			return "";
-		}
-	}
-	
-	private Pattern getLanguageKeyWordMatcher(String languageCode) {
-		try {
-			if (languageCode == null) {
-				languageCode = "en";
-			} else {
-				languageCode = languageCode.toLowerCase();
-			}
-			I18n i18n = new I18n(languageCode);
-			
-			StringBuilder sb = new StringBuilder();
-			sb.append("(?:");
-			String delim = "";
-		
-			for(String keyWord : i18n.getCodeKeywords()) {
-				sb.append(delim).append(keyWord);
-				delim = "|";
-			}
-		
-			return Pattern.compile((sb.append(") (.*)$").toString()));
-		} catch(NullPointerException e) {
-			e.printStackTrace();
-			return null;
-		} catch(PatternSyntaxException e) {
-			e.printStackTrace();
-			return null;
 		}
 	}
 }
